Move RegexExtenssions methods to RegExpUtilities
